fix channels, thingspeak

+++ src/thingspeak.rs
@@ -3,7 +3,7 @@ use crate::{Workspace};
 use std::fs;
 use serde::{Deserialize, Serialize};
 use chrono::{Utc};
-
+use glob::glob;
 // use serde::{Deserialize, Serialize};
 // use std::sync::mpsc::{self, channel};
 // use std::thread;
@@ -42,18 +42,18 @@ pub struct TSData {
     pub field6: Option<f32>,
     pub field7: Option<f32>,
     pub field8: Option<f32>,
-    pub latitude: Option<f32>,
-    pub longitude: Option<f32>,
-    pub elevation: Option<String>,
-    pub status: Option<String>
+    pub latitude: f32,
+    pub longitude: f32,
+    pub elevation: String,
+    pub status: String
 }
 
 impl TSData {
     pub fn new() -> TSData {
         let now = Utc::now();
-        log::info!("NEW TIMESTAMP:{}",now);
+        log::info!("NEW TIMESTAMP:{}",now.to_rfc3339());
         TSData{
-            created_at: format!("{}",now),
+            created_at: now.to_rfc3339(),
             field1: None,
             field2: None,
             field3: None,
@@ -62,10 +62,10 @@ impl TSData {
             field6: None,
             field7: None,
             field8: None,
-            latitude: None,
-            longitude: None,
-            elevation: None,
-            status: None
+            latitude: 52.4629112,
+            longitude: 13.4692855,
+            elevation: "".to_owned(),
+            status: "".to_owned()
         }
     }
 }
@@ -91,36 +91,57 @@ impl ThingSpeak {
         fs::read_to_string(self.path.join(CHANNEL)).unwrap_or(TSCHANNEL.to_owned())
     }
     /// change wkey
-    pub fn set_wkey(&self, key: &str) -> Result<()> {
-        fs::write(self.path.join(WKEY), key.as_bytes())?;
-        Ok(())
+    pub fn set_wkey(&self, key: &str) -> &Self {
+        if let Err(_e) = fs::write(self.path.join(WKEY), key.as_bytes()) {
+            log::error!("thingspeak change api write key failed");
+        }
+        self
     }
     /// change rkey
-    pub fn set_rkey(&self, key: &str) -> Result<()> {
-        fs::write(self.path.join(RKEY), key.as_bytes())?;
-        Ok(())
+    pub fn set_rkey(&self, key: &str) -> &Self {
+        if let Err(_) = fs::write(self.path.join(RKEY), key.as_bytes()) {
+            log::error!("thingspeak change api read key failed");
+        }
+        self
     }
     /// change channel
-    pub fn set_channel(&self, channel: &str) -> Result<()> {
-        fs::write(self.path.join(CHANNEL), channel.as_bytes())?;
-        Ok(())
+    pub fn set_channel(&self, channel: &str) -> &Self {
+        if let Err(_e) = fs::write(self.path.join(CHANNEL),channel.as_bytes()){
+            log::error!("thingspeak change channel failed");
+        }
+        self
     }
     pub fn update_url(&self) -> String {
         format!("https://api.thingspeak.com/channels/{}/bulk_update.json",self.channel().trim())
     }
     pub fn transmit(&self) -> Result<()> {
         let client = reqwest::blocking::Client::new();
-        for entry in fs::read_dir(&self.path)? {
-            let entry = entry?;
-            let path = entry.path();
-            let data:TSData = serde_json::from_str(&fs::read_to_string(&path)?)?;
-            let msg = TSMessage{
-                write_api_key: self.wkey(),
-                updates: vec![data],
-            };
-            let res = client.post(self.update_url().as_str()).json(&msg).send()?;
-            log::info!("Response Status: {}",res.status());
-            fs::remove_file(&path)?;
+        for entry in glob(format!("{}/*.json",self.path.display()).as_str()).unwrap() {
+            let path = entry.unwrap();
+            let data_str =  fs::read_to_string(&path)?;
+            match serde_json::from_str::<TSData>(&data_str) {
+                Ok(data) => {
+                    let msg = TSMessage{
+                        write_api_key: self.wkey(),
+                        updates: vec![data],
+                    };
+                    let res = client.post(self.update_url().as_str()).json(&msg).send()?;
+                    log::info!("Response Status: {}",res.status());
+                    if res.status().is_success(){
+                        fs::remove_file(&path)?;
+                    }else {
+                        break;
+                    }
+                },
+                Err(e) => {
+                    log::error!("TS {} entcode failed - {}",path.display(),e);
+                    fs::remove_file(&path)?;
+                }
+            }
+            
+            // let data_str = serde_json::to_string(&msg)?;
+            // println!("TS {} TRANSMIT:{}",path.display(),data_str);
+            //
         }
         Ok(())
     }
@@ -129,6 +150,7 @@ impl ThingSpeak {
         let path = self.path.join(format!("{}.json", now.format("%Y%m%dT%H%M%S")));
         let data_str = serde_json::to_string(&data)?;
         fs::write(path, data_str.as_bytes())?;
+        println!("TS PUBLISH:{}",data_str);
         self.transmit()
     }
 }
@@ -140,6 +162,8 @@ pub fn setup(ws: &Workspace) -> Result<ThingSpeak> {
     if !thingspeak.path.is_dir() {
         log::info!("Create new thingsspeak directory {}",thingspeak.path.as_path().display());
         fs::create_dir_all(&thingspeak.path)?;
+        thingspeak.set_wkey("3KBW4UF0N9KGJHU8").set_rkey("8MOUY0P3OFF9CECP").set_channel("1125745");
+
     }
     Ok(thingspeak)
 }
